finish test stuff (for now)
